+++ src/Fill/FillPolygon.py
@@ -23,8 +23,8 @@ def fill_polygon_fields(scene):
     scene.ui.polygon_def_str.setText(scene.project_data.items[ids[0]].definition_string())
     scene.ui.polygon_name.setText(polygon["id"])
 
-    scene.ui.polygon_line_width_slider.setValue(10.0 * polygon["line"]["line_width"])
-    scene.ui.polygon_line_width_spin.setValue(polygon["line"]["line_width"])
+    scene.ui.polygon_line_width_slider.setValue(int(10.0 * polygon["line"]["line_width"]))
+    scene.ui.polygon_line_width_spin.setValue(int(polygon["line"]["line_width"]))
 
     fill_colour(scene, polygon["fill"]["colour"], colours,
         scene.ui.polygon_marker_colour_name,
